feat(chore): change components to folder components
