+++ package.json
@@ -37,7 +37,7 @@
     "build": "next build && tsc --project tsconfig.server.json",
     "start": "cross-env NODE_ENV=production node dist/index.js",
     "prestart": "npm run build",
-    "relay": "./node_modules/.bin/relay-compiler --src ./pages --schema schema/schema.graphql --language typescript --artifactDirectory ./pages/__generated__",
+    "relay": "./node_modules/.bin/relay-compiler --src ./components --schema schema/schema.graphql --language typescript --artifactDirectory ./components/__generated__",
     "update-schema": "get-graphql-schema http://localhost:3000/graphql > schema/schema.graphql"
   },
   "devDependencies": {
